feat: add HOC appEntryPoint to mutualize logic between app an intents 🔨
